webui/requirements.txt
flask>=3.0.0
plexapi>=4.17.1
yt-dlp[default]>=2026.03.03
PyYAML>=6.0.3
requests
apscheduler>=3.11.0
